+++ lib/eve_online.rb
@@ -151,6 +151,7 @@ require 'eve_online/esi/models/region'
 require 'eve_online/esi/models/constellation'
 require 'eve_online/esi/models/category'
 require 'eve_online/esi/models/group'
+require 'eve_online/esi/models/graphic'
 
 require 'eve_online/esi/models/killmail_short'